# public/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.3
  title: The Booking System API
  description: The assignment for the Minut
paths:
...
  /v1/reservations/{id}:
    get:
      summary: Get a reservation by ID
      operationId: getReservation
      tags:
        - reservations
      parameters:
        - name: id
          in: path
          description: The reservation ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Reservation:
      type: object
      required:
        - id
        - propertyId
        - checkIn
        - checkOut
        - total
        - currency
      properties:
        id:
          type: string
          description: The reservation ID
        createdAt:
          type: string
          format: date-time
          description: The reservation creation date
        updatedAt:
          type: string
          format: date-time
          description: The reservation update date
        propertyId:
          type: string
          description: The property ID
        guestId:
          type: string
          description: The guest ID
        checkIn:
          type: string
          format: date-time
          description: The reservation check-in date
        checkOut:
          type: string
          format: date-time
          description: The reservation check-out date
        total:
          type: number
          format: float
          description: The reservation total
        currency:
          type: string
          description: The reservation currency
        guests:
          type: array
          items:
            type: object
            required:
              - id
              - name
              - email
              - phone
            properties:
              id:
                type: string
                description: The guest ID
              name:
                type: string
                description: The guest name
              email:
                type: string
                format: email
                description: The guest email
              phone:
                type: string
                description: The guest phone
...